Use single linkingConfig for deep links
