+++ src/Navigator/NavigationContainer/AppNavigator/index.js
@@ -1,3 +1,7 @@
-import ResponsiveNavigator from './ResponsiveNavigator';
+import React from 'react';
+import AppNavigator from './AppNavigator';
 
-export default ResponsiveNavigator;
+export default props => (
+    // eslint-disable-next-line react/jsx-props-no-spreading
+    <AppNavigator responsive {...props} />
+);
